FE: Refactor Game file
