+++ PandaBattleship.FE/src/components/Game.jsx
@@ -1,13 +1,9 @@
 ﻿import {useState, useEffect} from 'react';
 import { SHIP_LAYOUTS } from '../constants/shipLayouts';
-
-const GRID_SIZE = 10;
+import { createEmptyGrid, findSunkenShip, markSunkShipOnGrid, GRID_SIZE } from '../utils/gameHelpers';
+import Grid from './Grid';
 
 const Game = () => {
-    // Helper to create an empty grid
-    const createEmptyGrid = () => Array(GRID_SIZE)
-        .fill(null)
-        .map(() => Array(GRID_SIZE).fill(null));
 
     const [playerGrid, setPlayerGrid] = useState(createEmptyGrid());
     const [enemyGrid, setEnemyGrid] = useState(createEmptyGrid());
@@ -34,55 +30,6 @@ const Game = () => {
         setEnemyShipLayout(SHIP_LAYOUTS[enemyLayoutIdx]);
     }, []);
 
-    const getSurroundingCells = (shipCoords) => {
-        const surrounding = new Set();
-        shipCoords.forEach(([r, c]) => {
-            for (let dr = -1; dr <= 1; dr++) {
-                for (let dc = -1; dc <= 1; dc++) {
-                    const nr = r + dr;
-                    const nc = c + dc;
-                    if (nr >= 0 && nr < GRID_SIZE && nc >= 0 && nc < GRID_SIZE) {
-                        // Don't include the ship's own cells
-                        if (!shipCoords.some(([sr, sc]) => sr === nr && sc === nc)) {
-                            surrounding.add(`${nr},${nc}`);
-                        }
-                    }
-                }
-            }
-        });
-        return surrounding;
-    };
-
-    const findSunkenShip = (grid, layout) => {
-        if (!layout) return null;
-
-        for (const ship of layout.ships) {
-            const allHit = ship.coords.every(([r, c]) => grid[r][c] === 'hit' || grid[r][c] === 'sunk');
-            const isAlreadyMarkedSunk = ship.coords.every(([r, c]) => grid[r][c] === 'sunk');
-            if (allHit && !isAlreadyMarkedSunk) {
-                return ship;
-            }
-        }
-        return null;
-    };
-
-    const markSunkShipOnGrid = (grid, sunkenShip) => {
-        const newGrid = grid.map(r => [...r]);
-        // Mark ship cells as 'sunk'
-        sunkenShip.coords.forEach(([r, c]) => {
-            newGrid[r][c] = 'sunk';
-        });
-
-        // Mark surrounding cells as 'blocked'
-        const surrounding = getSurroundingCells(sunkenShip.coords);
-        surrounding.forEach(coords => {
-            const [r, c] = coords.split(',').map(Number);
-            if (newGrid[r][c] === null) {
-                newGrid[r][c] = 'blocked';
-            }
-        });
-        return newGrid;
-    };
 
     const performAiTurn = async () => {
         let continueAi = true;
@@ -163,30 +110,6 @@ const Game = () => {
         }
     };
 
-
-    const renderGrid = (grid, onCellClick, isPlayerGrid = false) => {
-        return (
-            <div className="inline-grid grid-cols-10 gap-0 border border-gray-500">
-                {grid.map((rowArr, rowIndex) =>
-                    rowArr.map((cell, colIndex) => (
-                        <span
-                            key={`${rowIndex}-${colIndex}`}
-                            className={`w-8 h-8 sm:w-10 sm:h-10 border border-gray-600 flex items-center justify-center cursor-pointer hover:bg-gray-700 transition-colors ${
-                                cell === 'sunk' ? 'bg-orange-500' :
-                                    cell === 'hit' ? 'text-red-500 font-bold' :
-                                        cell === 'miss' || cell === 'blocked' ? 'text-blue-400' :
-                                            (isPlayerGrid && cell === 'ship') ? 'bg-gray-500' : ''
-                            }`}
-                            onClick={() => onCellClick && onCellClick(rowIndex, colIndex)}
-                        >
-                            {cell === 'hit' ? '🔥' : (cell === 'miss' || cell === 'blocked') ? 'O' : ''}
-                        </span>
-                    ))
-                )}
-            </div>
-        );
-    };
-
     return (
         <div className="flex flex-col items-center gap-8 p-4">
             <div className={`text-2xl font-bold p-4 rounded-lg transition-colors ${isPlayerTurn ? 'text-green-400 bg-green-900/20' : 'text-red-400 bg-red-900/20'}`}>
@@ -195,12 +118,12 @@ const Game = () => {
 
             <div className="flex flex-col items-center">
                 <h2 className="text-xl mb-2 text-red-400">Enemy Waters</h2>
-                {renderGrid(enemyGrid, handleEnemyCellClick)}
+                <Grid grid={enemyGrid} onCellClick={handleEnemyCellClick} />
             </div>
 
             <div className="flex flex-col items-center">
                 <h2 className="text-xl mb-2 text-green-400">Your Fleet</h2>
-                {renderGrid(playerGrid, null, true)}
+                <Grid grid={playerGrid} onCellClick={null} isPlayerGrid={true} />
             </div>
 
             <div className="text-xs text-gray-500 max-h-24 overflow-y-auto">
