+++ PandaBattleship.FE/src/components/Grid.jsx
@@ -1,0 +1,24 @@
+﻿const Grid = ({ grid, onCellClick, isPlayerGrid = false }) => {
+    return (
+        <div className="inline-grid grid-cols-10 gap-0 border border-gray-500">
+            {grid.map((rowArr, rowIndex) =>
+                rowArr.map((cell, colIndex) => (
+                    <span
+                        key={`${rowIndex}-${colIndex}`}
+                        className={`w-8 h-8 sm:w-10 sm:h-10 border border-gray-600 flex items-center justify-center cursor-pointer hover:bg-gray-700 transition-colors ${
+                            cell === 'sunk' ? 'bg-orange-500' :
+                                cell === 'miss' || cell === 'blocked' ? 'text-blue-400' :
+                                    (isPlayerGrid && cell === 'hit') ? 'bg-gray-500':
+                                        (isPlayerGrid && cell === 'ship') ? 'bg-gray-500' : ''
+                        }`}
+                        onClick={() => onCellClick && onCellClick(rowIndex, colIndex)}
+                    >
+                        {cell === 'hit' ? '🔥' : (cell === 'miss' || cell === 'blocked') ? 'O' : ''}
+                    </span>
+                ))
+            )}
+        </div>
+    );
+};
+
+export default Grid;
